+++ lib/rubocop/cop/style/empty_literal.rb
@@ -15,16 +15,15 @@ module RuboCop
         def_node_matcher :array_node, '(send (const nil :Array) :new)'
         def_node_matcher :hash_node, '(send (const nil :Hash) :new)'
         def_node_matcher :str_node, '(send (const nil :String) :new)'
+        def_node_matcher :array_with_block,
+                         '(block (send (const nil :Array) :new) args _)'
+        def_node_matcher :hash_with_block,
+                         '(block (send (const nil :Hash) :new) args _)'
 
         def on_send(node)
-          array_node(node) { add_offense(node, :expression, ARR_MSG) }
+          add_offense(node, :expression, ARR_MSG) if offense_array_node?(node)
 
-          hash_node(node) do
-            # If Hash.new takes a block, it can't be changed to {}.
-            return if node.parent && node.parent.block_type?
-
-            add_offense(node, :expression, HASH_MSG)
-          end
+          add_offense(node, :expression, HASH_MSG) if offense_hash_node?(node)
 
           str_node(node) do
             return if frozen_string_literals_enabled?
@@ -77,12 +76,21 @@ module RuboCop
           end
         end
 
+        def offense_array_node?(node)
+          array_node(node) && !array_with_block(node.parent)
+        end
+
+        def offense_hash_node?(node)
+          # If Hash.new takes a block, it can't be changed to {}.
+          hash_node(node) && !hash_with_block(node.parent)
+        end
+
         def correction(node)
-          if array_node(node)
+          if offense_array_node?(node)
             '[]'
           elsif str_node(node)
             preferred_string_literal
-          elsif hash_node(node)
+          elsif offense_hash_node?(node)
             if first_argument_unparenthesized?(node)
               # `some_method {}` is not same as `some_method Hash.new`
               # because the braces are interpreted as a block. We will have